SRU::Client v0.02 with improved tests and documentation
As the last version of SRU was 0.99, I have increased
the version number to 1.00.  I've never been a part of
a major version number change before.  It's exciting!

Trying to fit in better with the SRU bundle, I have removed the
requirement for Moose, LWP::Simple and Carp from Makefile.PL and
put a skip_all in the test file (renamed from client.t to
sru_client.t to put it further down the test order because it
takes time for the responses to return) if it fails to find
any of the required modules, as it does in cgi_app.t

I uncovered a problem with having HTTP_PROXY set and documented
it, although having the user uncomment a line in sru_client.t is
not optimal.